requirements.txt
streamlit
python-dotenv
google-generativeai
openai
groq
